+++ Sources/DTModelStorage/EventReactions.swift
@@ -32,9 +32,6 @@ import UIKit
 /// Data holder for reaction
 open class EventReaction {
     
-    /// view -> model mapping of this reaction
-    public let viewModelMapping: ViewModelMapping
-    
     /// 3 arguments reaction block with all arguments type-erased.
     open var reaction : ((Any, Any, Any) -> Any)?
     
@@ -42,34 +39,27 @@ open class EventReaction {
     public let methodSignature: String
     
     /// Creates reaction with `signature`.
-    public init<T: ModelTransfer>(signature: String, viewType: ViewType, viewClass: T.Type) {
+    public init<View, ModelType, ReturnType>(_ viewType: View.Type,
+                                            _ modelType: ModelType.Type,
+                                            signature: String,
+                                            _ block: @escaping (View, ModelType, IndexPath) -> ReturnType) {
         self.methodSignature = signature
-        viewModelMapping = ViewModelMapping.eventsViewMapping(viewType: viewType, viewClass: T.self)
-    }
-    
-    /// Creates reaction with `signature`, `viewType` and `modelType`.
-    public init<T>(signature: String, viewType: ViewType, modelType: T.Type) {
-        self.methodSignature = signature
-        viewModelMapping = ViewModelMapping.eventsModelMapping(viewType: viewType, modelClass: T.self)
-    }
-    
-    /// Creates reaction with type-erased T and T.ModelType into Any types.
-    open func makeReaction<T: ModelTransfer, U>(_ block: @escaping (T, T.ModelType, IndexPath) -> U) {
-        reaction = { cell, model, indexPath in
-            guard let model = model as? T.ModelType,
+        reaction = { view, model, indexPath in
+            guard let model = model as? ModelType,
                 let indexPath = indexPath as? IndexPath,
-                let cell = cell as? T
+                let view = view as? View
                 else {
                     return 0
             }
-            return block(cell, model, indexPath)
+            return block(view, model, indexPath)
         }
     }
     
-    /// Creates reaction with type-erased T into Any type.
-    open func makeReaction<T, U>(_ block: @escaping (T, IndexPath) -> U) {
+    /// Creates reaction with `signature`, `viewType` and `modelType`.
+    public init<Argument, ReturnType>(_ modelType: Argument.Type, signature: String, _ block: @escaping (Argument, IndexPath) -> ReturnType) {
+        self.methodSignature = signature
         reaction = { cell, model, indexPath in
-            guard let model = model as? T,
+            guard let model = model as? Argument,
                 let indexPath = indexPath as? IndexPath else {
                     return 0
             }
@@ -88,26 +78,37 @@ open class FourArgumentsEventReaction: EventReaction {
     
     /// Type-erased reaction with 4 arguments
     open var reaction4Arguments : ((Any, Any, Any, Any) -> Any)?
+   
+    @available(*, unavailable)
+    public override init<Argument, ReturnType>(_ modelType: Argument.Type, signature: String, _ block: @escaping (Argument, IndexPath) -> ReturnType) {
+        super.init(modelType, signature: signature, block)
+    }
     
-    /// Makes 4 argument reaction based on `View`.
-    open func make4ArgumentsReaction<View:ModelTransfer, Argument, ReturnType>(_ block: @escaping (Argument, View, View.ModelType, IndexPath) -> ReturnType) {
+    @available(*, unavailable)
+    public override init<View, Argument, ReturnType>(_ viewType: View.Type, _ modelType: Argument.Type, signature: String, _ block: @escaping (View, Argument, IndexPath) -> ReturnType) {
+        super.init(viewType, modelType, signature: signature, block)
+    }
+    
+    @available(*, unavailable)
+    open override func performWithArguments(_ arguments: (Any, Any, Any)) -> Any {
+        fatalError("This method should not be called. Please call 4 argument version of this method")
+    }
+    
+    public init<View, ModelType, Argument, ReturnType>(_ viewType: View.Type,
+                                                         modelType: ModelType.Type,
+                                                         argument: Argument.Type,
+                                                         signature: String,
+                                                         _ block: @escaping (Argument, View, ModelType, IndexPath) -> ReturnType) {
+        super.init(viewType, modelType, signature: signature) { _, _, _ in
+            fatalError("This closure should not be called by FourArgumentsEventReaction")
+        }
         reaction4Arguments = { argument, view, model, indexPath in
-            guard let model = model as? View.ModelType,
+            guard let model = model as? ModelType,
                 let indexPath = indexPath as? IndexPath,
                 let argument = argument as? Argument,
                 let view = view as? View else { return 0 }
             return block(argument, view, model, indexPath)
         }
-    }
-    
-    /// Creates FourArgumentsEventReaction for `viewClass`
-    public override init<T: ModelTransfer>(signature: String, viewType: ViewType, viewClass: T.Type) {
-        super.init(signature: signature, viewType: viewType, viewClass: viewClass)
-    }
-    
-    /// Creates FourArgumentsEventReaction for `modelType`
-    public override init<T>(signature: String, viewType: ViewType, modelType: T.Type) {
-        super.init(signature: signature, viewType: viewType, modelType: modelType)
     }
     
     /// Performs reaction with `arguments`.
@@ -122,10 +123,31 @@ open class FiveArgumentsEventReaction: EventReaction {
     /// Type-erased reaction with 5 arguments
     open var reaction5Arguments : ((Any, Any, Any, Any, Any) -> Any)?
     
-    /// Makes 5 argument reaction based on `View`.
-    open func make5ArgumentsReaction<View:ModelTransfer, ArgumentOne, ArgumentTwo, ReturnType>(_ block: @escaping (ArgumentOne, ArgumentTwo, View, View.ModelType, IndexPath) -> ReturnType) {
+    @available(*, unavailable)
+    public override init<Argument, ReturnType>(_ modelType: Argument.Type, signature: String, _ block: @escaping (Argument, IndexPath) -> ReturnType) {
+        super.init(modelType, signature: signature, block)
+    }
+    
+    @available(*, unavailable)
+    public override init<View, Argument, ReturnType>(_ viewType: View.Type, _ modelType: Argument.Type, signature: String, _ block: @escaping (View, Argument, IndexPath) -> ReturnType) {
+        super.init(viewType, modelType, signature: signature, block)
+    }
+    
+    @available(*, unavailable)
+    open override func performWithArguments(_ arguments: (Any, Any, Any)) -> Any {
+        fatalError("This method should not be called. Please call 4 argument version of this method")
+    }
+    
+    public init<View, ModelType, ArgumentOne, ArgumentTwo, ReturnType>(_ viewType: View.Type,
+                                                                       modelType: ModelType.Type,
+                                                                       argumentOne: ArgumentOne.Type,
+                                                                       argumentTwo: ArgumentTwo.Type,
+                                                                       signature: String, _ block: @escaping (ArgumentOne, ArgumentTwo, View, ModelType, IndexPath) -> ReturnType) {
+        super.init(viewType, modelType, signature: signature) { _, _, _ in
+            fatalError("This closure should not be called by FiveArgumentsEventReaction")
+        }
         reaction5Arguments = { argumentOne, argumentTwo, view, model, indexPath in
-            guard let model = model as? View.ModelType,
+            guard let model = model as? ModelType,
                 let indexPath = indexPath as? IndexPath,
                 let argument1 = argumentOne as? ArgumentOne,
                 let argument2 = argumentTwo as? ArgumentTwo,
@@ -134,40 +156,34 @@ open class FiveArgumentsEventReaction: EventReaction {
         }
     }
     
-    /// Creates FiveArgumentsEventReaction for `viewClass`
-    public override init<T: ModelTransfer>(signature: String, viewType: ViewType, viewClass: T.Type) {
-        super.init(signature: signature, viewType: viewType, viewClass: viewClass)
-    }
-    
-    /// Creates FiveArgumentsEventReaction for `modelType`
-    public override init<T>(signature: String, viewType: ViewType, modelType: T.Type) {
-        super.init(signature: signature, viewType: viewType, modelType: modelType)
-    }
-    
     /// Performs reaction with `arguments`.
     open func performWithArguments(_ arguments: (Any, Any, Any, Any, Any)) -> Any {
         return reaction5Arguments?(arguments.0, arguments.1, arguments.2, arguments.3, arguments.4) ?? 0
     }
 }
 
-extension Sequence where Self.Iterator.Element: EventReaction {
+extension Sequence where Self.Iterator.Element: ViewModelMappingProtocol {
     /// Searches for reaction using specified parameters.
     public func reaction(of type: ViewType,
                          signature: String,
                          forModel model: Any,
+                         at indexPath: IndexPath,
                          view: UIView?) -> EventReaction? {
-        return first(where: { reaction in
-            guard let unwrappedModel = RuntimeHelper.recursivelyUnwrapAnyValue(model) else { return false }
-            return reaction.viewModelMapping.viewType == type &&
-                reaction.viewModelMapping.modelTypeCheckingBlock(unwrappedModel) &&
-                view?.isKind(of: reaction.viewModelMapping.viewClass) ?? true &&
-                reaction.methodSignature == signature
-        })
+        guard let unwrappedModel = RuntimeHelper.recursivelyUnwrapAnyValue(model) else { return nil }
+        return filter { mapping in
+            // Find all compatible mappings
+            mapping.modelTypeCheckingBlock(unwrappedModel) &&
+            mapping.viewType == type &&
+            (view?.isKind(of: mapping.viewClass) ?? true) &&
+            mapping.condition.isCompatible(with: indexPath, model: unwrappedModel) &&
+                mapping.reactions.contains { $0.methodSignature == signature }
+        }
+        .first?.reactions.first(where: { $0.methodSignature == signature })
     }
     
     /// Performs reaction of `type`, `signature`, with `view`, `model` in `location`.
-    public func performReaction(of type: ViewType, signature: String, view: Any?, model: Any, location: Any) -> Any {
-        guard let reaction = reaction(of: type, signature: signature, forModel: model, view: view as? UIView) else {
+    public func performReaction(of type: ViewType, signature: String, view: Any?, model: Any, location: IndexPath) -> Any {
+        guard let reaction = reaction(of: type, signature: signature, forModel: model, at: location, view: view as? UIView) else {
             return 0
         }
         return reaction.performWithArguments((view ?? 0, model, location))
@@ -175,14 +191,14 @@ extension Sequence where Self.Iterator.Element: EventReaction {
     
     //swiftlint:disable function_parameter_count
     /// Performs reaction of `type`, `signature`, with `argument`, `view`, `model` in `location`.
-    public func perform4ArgumentsReaction(of type: ViewType, signature: String, argument: Any, view: Any?, model: Any, location: Any) -> Any {
-        guard let reaction = reaction(of: type, signature: signature, forModel: model, view: view as? UIView) as? FourArgumentsEventReaction else { return 0 }
+    public func perform4ArgumentsReaction(of type: ViewType, signature: String, argument: Any, view: Any?, model: Any, location: IndexPath) -> Any {
+        guard let reaction = reaction(of: type, signature: signature, forModel: model, at: location, view: view as? UIView) as? FourArgumentsEventReaction else { return 0 }
         return reaction.performWithArguments((argument, view ?? 0, model, location))
     }
     
     /// Performs reaction of `type`, `signature`, with `firstArgument`, `secondArgument`, `view`, `model` in `location`.
-    public func perform5ArgumentsReaction(of type: ViewType, signature: String, firstArgument: Any, secondArgument: Any, view: Any?, model: Any, location: Any) -> Any {
-        guard let reaction = reaction(of: type, signature: signature, forModel: model, view: view as? UIView) as? FiveArgumentsEventReaction else { return 0 }
+    public func perform5ArgumentsReaction(of type: ViewType, signature: String, firstArgument: Any, secondArgument: Any, view: Any?, model: Any, location: IndexPath) -> Any {
+        guard let reaction = reaction(of: type, signature: signature, forModel: model, at: location, view: view as? UIView) as? FiveArgumentsEventReaction else { return 0 }
         return reaction.performWithArguments((firstArgument, secondArgument, view ?? 0, model, location))
     }
 }