+++ Sources/Tests/Specs/UIReactionsTestCase.swift
@@ -12,11 +12,16 @@ import DTModelStorage
 
 class UIReactionsTestCase: XCTestCase {
     
-    var reactions : [EventReaction]!
+    var mapping : ViewModelMapping<CollectionCell, Int>!
     
     override func setUp() {
         super.setUp()
-        reactions = []
+        mapping = ViewModelMapping<CollectionCell, Int>(cellConfiguration: { _,_,_ in }, mapping: { mapping in })
+    }
+    
+    override func tearDown() {
+        super.tearDown()
+        mapping.reactions.removeAll()
     }
     
     func testReactionTypeEquatable() {
@@ -29,71 +34,40 @@ class UIReactionsTestCase: XCTestCase {
     }
     
     func testReactionsAreSearchable() {
-        let reaction = EventReaction(signature: "foo", viewType: .cell, modelType: Int.self)
-        reaction.makeReaction(makeCellBlock({  }, cell: TableCell(), returnValue: 5))
-        reactions.append(reaction)
-        
-        let foundReaction = reactions.reaction(of: .cell, signature: "foo", forModel: 5, view: nil)
+        mapping.reactions.append(EventReaction(CollectionCell.self, Int.self, signature: "foo") { _,_,_ in })
+        let foundReaction = [mapping].reaction(of: .cell, signature: "foo", forModel: 5, at: indexPath(0, 0), view: nil)
         XCTAssertNotNil(foundReaction)
     }
     
     func testReactionsForOptionalModelsAreSearchable() {
-        let reaction = EventReaction(signature: "foo", viewType: .cell, modelType: Int.self)
-        reaction.makeReaction(makeCellBlock({  }, cell: TableCell(), returnValue: 5))
-        reactions.append(reaction)
+        mapping.reactions.append(EventReaction(CollectionCell.self, Int.self, signature: "foo") { _,_,_ in })
         
         let nilModel: Int? = 5
         
-        let foundReaction = reactions.reaction(of: .cell, signature: "foo", forModel: nilModel as Any, view: nil)
+        let foundReaction = [mapping].reaction(of: .cell, signature: "foo", forModel: nilModel as Any, at: indexPath(0, 0), view: nil)
         XCTAssertNotNil(foundReaction)
     }
     
-    func makeCellBlock<T, U>(_ block: @escaping () -> Void, cell: T, returnValue: U) -> (T?, T.ModelType, IndexPath) -> U
-        where T: ModelTransfer
-    {
-        return { one, two, three in
-            block()
-            return returnValue
-        }
-    }
-    
-    func makeSupplementaryBlock<T, U>(_ block: @escaping () -> Void, cell: T, returnValue: U) -> (T?, T.ModelType, IndexPath) -> U where T: ModelTransfer {
-        return { one, two, three in
-            block()
-            return returnValue
-        }
-    }
-    
     func testCellReactionIsExecutable() {
-        let reaction = EventReaction(signature: "foo", viewType: .cell, modelType: Int.self)
         let exp = expectation(description: "executeCell")
-        reaction.makeReaction(makeCellBlock({
+        let reaction = EventReaction(CollectionCell.self, Int.self, signature: "foo") { _,_,_ -> Int in
             exp.fulfill()
-            }, cell: TableCell(), returnValue: 3))
-        let result = reaction.performWithArguments((TableCell(), 5, indexPath(0, 0)))
-        waitForExpectations(timeout: 1, handler: nil)
-        XCTAssertEqual(result as? Int, 3)
-    }
-    
-    func testSupplementaryReactionIsExecutable() {
-        let reaction = EventReaction(signature: "foo", viewType: .supplementaryView(kind: "Foo"), modelType: Int.self)
-        let exp = expectation(description: "executeCell")
-        reaction.makeReaction(makeSupplementaryBlock({
-            exp.fulfill()
-            }, cell: TableCell(), returnValue: 3))
-        let result = reaction.performWithArguments((TableCell(), 5, indexPath(0, 5)))
+            return 3
+        }
+        mapping.reactions.append(reaction)
+        let result = reaction.performWithArguments((CollectionCell(), 5, indexPath(0, 0)))
         waitForExpectations(timeout: 1, handler: nil)
         XCTAssertEqual(result as? Int, 3)
     }
     
     func testReactionOfTypeIsPerformable() {
-        let reaction = EventReaction(signature: "foo", viewType: .cell, modelType: Int.self)
         let exp = expectation(description: "executeCell")
-        reaction.makeReaction(makeCellBlock({
+        let reaction = EventReaction(CollectionCell.self, Int.self, signature: "foo") { _,_,_ -> Int in
             exp.fulfill()
-            }, cell: TableCell(), returnValue: 3))
-        reactions.append(reaction)
-        let result = reactions.performReaction(of: .cell, signature: "foo", view: TableCell(), model: 5, location: indexPath(0, 0))
+            return 3
+        }
+        mapping.reactions.append(reaction)
+        let result = [mapping].performReaction(of: .cell, signature: "foo", view: CollectionCell(), model: 5, location: indexPath(0, 0))
         waitForExpectations(timeout: 1, handler: nil)
         XCTAssertEqual(result as? Int, 3)
     }